+++ apps/de.hub.clickwatch.apps.recorder.examples/src/de/hub/clickwatch/recorder/examples/lib/RemoveOffset.java
@@ -7,7 +7,7 @@ public class RemoveOffset {
 		mavg = new MovingAvg(windowSize);
 	}
 	
-	public int filter(int value) {
+	public double filter(double value) {
 		return value - mavg.filter(value);
 	}
 }
